# Avoid using Module#<= in value_for