@rafaelfranca 2022-09-01
Significant? Isn't this in the standard deviation of 9%?