@jhawthorn 2022-09-01
It's significant (though the benchmark there shows a small difference).

On my desktop machine, which has a much better stddev

```
       Comment.where     39.193k (± 0.2%) i/s -    198.594k in   5.067145s
```

after:

```
       Comment.where     40.692k (± 0.2%) i/s -    205.050k in   5.039054s
```

We can make this have more of an impact by including a bunch of modules onto `Comment` (in currently released Rubies `<=` needs to scan through all included Modules, though this is being fixed in Ruby 3.2)

We could also change the benchmark to test `Comment.current_scope` directly

Before: `Comment.current_scope     1.336M (± 0.9%) i/s -      6.683M in   5.003522s`
Before: `Comment.current_scope     2.060M (± 2.7%) i/s -     10.478M in   5.089445s`