@rafaelfranca 2022-09-01
The `current_scope` benchmark shows enough significant difference in performance to justify this.